+++ application/home_works_examples/read_text.py
@@ -1,0 +1,11 @@
+from application.config import FILES_INPUT_FILES
+
+
+def read_text():
+    data = FILES_INPUT_FILES
+    with open(data) as a:
+        reading = a.read()
+    return reading
+
+
+print(read_text())
